support the crazy nonsense of a NATIONAL charset
before the days of the sanity and light that is utf8, there were people
who thought it would be good to have a character set that changed
depending on what country you might be in.  40 years later, I spend 3
hours of wasted time on that mistake.  Interesting thing, time and
history.